implemented controllers, fixed dto

+++ MoneyTransferSystem/src/main/java/com/fidelity/mts/application/dto/TransferRequest.java
@@ -1,8 +1,9 @@
 package com.fidelity.mts.application.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record TransferRequest(Integer fromAccountId,Integer toAccountId,
-		Long amount, UUID idempotencyKey) {
+		BigDecimal amount, UUID idempotencyKey) {
 
 }
